+++ NEWS.md
@@ -1,6 +1,8 @@
 # TreeTools 1.3.1.9003 (development)
 
 - Add `replaceAll` option to `ArtificialExtinction()`.
+- Write different character types to TNT file with 
+  `WriteTntCharacters(types = ...)`.
 
 # TreeTools 1.3.1.9002 (development)